nemo/requirements.txt
langchain==0.0.167
nemoguardrails==0.3.0
openai==0.27.2
python-dotenv==1.0.0
requests==2.31.0